Control size of pairwise result file by OabaSettings.getMaxMatches()

+++ 2.7.x/choicemaker-cm/choicemaker-j2ee/com.choicemaker.cm.batch.ejb/src/main/java/com/choicemaker/cm/batch/ejb/BatchJobFileUtils.java
@@ -32,7 +32,7 @@ public class BatchJobFileUtils {
 
 	public static final String FMT = "00000";
 
-	public static final long MAX_FILE_SIZE = 1500000000;
+	public static final int MAX_FILE_SIZE = 1500000000;
 
 	public static final int ROLLOVER = 100000;
